pyproject.toml
[project]
name = "mustang_sage"
version = "0.1.0"
description = "The Mustang Sage - Teams AI Co-pilot"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "botbuilder-core>=4.15.0",
    "botbuilder-integration-aiohttp>=4.15.0",
    "langgraph>=0.0.10",
    "langchain>=0.1.0",
    "tenacity>=8.2.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "pytest>=9.0.3",
    "pydantic-settings>=2.13.1",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.0,<4.0.0",
    "pytest-cov>=4.1.0",
    "jsonschema>=4.21.1",
    "schedule>=1.2.1",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
]
